submit

+++ baiTap/Bai1.c
@@ -1,31 +1,30 @@
 #include <stdio.h>
 
-long tong(int n) {
+int tong(int n) {
     if (n == 0)
         return 0;
     return n + tong(n - 1);
 }
 
-long tich(int n) {
+int tich(int n) {
     if (n == 1)
         return 1;
     return n * tich(n - 1);
 }
 
-float luyThua(float a, int n) {
+double luyThua(double a, int n) {
     if (n == 0) 
         return 1;
     return a * luyThua(a, n - 1);
 }
 
-
 int main() {
-    float a;
+    double a;
     int n;
     printf("Nhap n = ");
     scanf("%d", &n);
     printf("Nhap a = ");
-    scanf("%f", &a);
+    scanf("%lf", &a);
 
     while(n < 0) {
         printf("Du lieu khong hop le. Nhap lai n = ");
@@ -34,7 +33,7 @@ int main() {
     
     printf("S(%d) = %d", n, tong(n));
     printf("\nP(%d) = %d", n, tich(n));
-    printf("\nA(%.2f, %d) = %.2f", a, n, luyThua(a, n));
+    printf("\nA(%.0lf, %d) = %.0lf", a, n, luyThua(a, n));
 
     return 0;
 }
